Libtock-c is PR ready
