+++ examples/tests/qdec_callback/Makefile
@@ -1,7 +1,7 @@
 # Makefile for user application
 
 # Specify this directory relative to the current application.
-TOCK_USERLAND_BASE_DIR = ../..
+TOCK_USERLAND_BASE_DIR = ../../..
 
 # Which files to compile.
 C_SRCS := $(wildcard *.c)
